First commit!
The “Batalha” game is not yet fully implemented; it depends on a
“Personagem” class, that is (I suppose) fully implemented, therefore I
will finish the game in a minute of my free time.
And yeah, I know, the code is in portuguese but the commit comments are
in english, but this is for learning! And my class is in pt-br, so ok.